change: rename startserver package to controlserver
